requirements.txt
openai
python-dotenv
streamlit
streamlit-chat
sqlalchemy
pandas
requests
